SK-24 Care Worker Appointment Crud
Progress Save

Pending Issues:
The edit still doesn't preserve older appointments that should be preserved. To check for missing functions that I might have accidentally deleted, I think the creation of appointment in their place is not working this cannot be a display problem anymore.

The cancel appointment still doesn't work here, it should not ask for occurrence if it is not recurring.

Pending Feature:
Now, the thing for care workers is that they can be the ones to decide the time of a flexible appointment. My problem is that if it is recurring, they should be able to choose whether to just assign a specific time to that occurrence, or also to all the next recurring ones through edit.